+++ src/ReactiveDomain.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 18
-VisualStudioVersion = 18.4.11612.150 stable
+VisualStudioVersion = 18.4.11612.150
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ReactiveDomain.Messaging", "ReactiveDomain.Messaging\ReactiveDomain.Messaging.csproj", "{3BAF7404-334D-4FD3-97A5-4D15F3049334}"
 EndProject
@@ -54,6 +54,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ReactiveDomain.IdentityStorage.Tests", "ReactiveDomain.IdentityStorage.Tests\ReactiveDomain.IdentityStorage.Tests.csproj", "{46F5E314-9AAF-419D-A261-940D55CD0FC9}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ReactiveDomain.PolicyTool", "ReactiveDomain.PolicyTool\ReactiveDomain.PolicyTool.csproj", "{44879F67-D0AD-4D81-B680-CB01FBC5B75F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReactiveDomain.Testing.Tests", "ReactiveDomain.Testing.Tests\ReactiveDomain.Testing.Tests.csproj", "{C6DF94CA-6BAC-4A08-A593-D8404CF3D55B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -123,6 +125,10 @@ Global
 		{46F5E314-9AAF-419D-A261-940D55CD0FC9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{44879F67-D0AD-4D81-B680-CB01FBC5B75F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{44879F67-D0AD-4D81-B680-CB01FBC5B75F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C6DF94CA-6BAC-4A08-A593-D8404CF3D55B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C6DF94CA-6BAC-4A08-A593-D8404CF3D55B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C6DF94CA-6BAC-4A08-A593-D8404CF3D55B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C6DF94CA-6BAC-4A08-A593-D8404CF3D55B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
